Implement changes to finalize structure Implement changes to finalize structure
