+++ braidpool-primitives/src/lib.rs
@@ -1,3 +1,5 @@
 pub mod bead;
 pub mod braid;
+pub mod committed_metadata;
+pub mod uncommitted_metadata;
 pub mod utils;
