+++ braidpool-primitives/src/uncommitted_metadata/mod.rs
@@ -1,0 +1,73 @@
+use ::serde::{Deserialize, Serialize};
+use bitcoin::absolute::Time;
+use bitcoin::consensus::Error;
+use bitcoin::consensus::encode::Decodable;
+use bitcoin::consensus::encode::Encodable;
+use bitcoin::ecdsa::Signature;
+use bitcoin::io::{self, BufRead, Write};
+use core::str::FromStr;
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq, Default)]
+pub struct TimeVec(pub Vec<Time>);
+
+impl Encodable for TimeVec {
+    fn consensus_encode<W: Write + ?Sized>(&self, w: &mut W) -> Result<usize, io::Error> {
+        let mut len = 0;
+        // Encode the length for deterministic encoding
+        len += (self.0.len() as u64).consensus_encode(w)?;
+        for time in &self.0 {
+            len += time.to_consensus_u32().consensus_encode(w)?;
+        }
+        Ok(len)
+    }
+}
+
+impl Decodable for TimeVec {
+    fn consensus_decode<R: BufRead + ?Sized>(r: &mut R) -> Result<Self, Error> {
+        let len = u64::consensus_decode(r)?;
+        let mut vec = Vec::with_capacity(len as usize);
+        for _ in 0..len {
+            let time_u32 = u32::consensus_decode(r)?;
+            let time = Time::from_consensus(time_u32).unwrap();
+            vec.push(time);
+        }
+        Ok(TimeVec(vec))
+    }
+}
+
+#[derive(Clone, Debug, Serialize, PartialEq)]
+pub struct UnCommittedMetadata {
+    pub extra_nonce: i32,
+    pub broadcast_timestamp: Time,
+    pub signature: Signature,
+    pub parent_bead_timestamps: TimeVec,
+}
+
+impl Encodable for UnCommittedMetadata {
+    fn consensus_encode<W: Write + ?Sized>(&self, w: &mut W) -> Result<usize, io::Error> {
+        let mut len = 0;
+        len += self.extra_nonce.consensus_encode(w)?;
+        len += self
+            .broadcast_timestamp
+            .to_consensus_u32()
+            .consensus_encode(w)?;
+        len += self.signature.to_string().consensus_encode(w)?;
+        len += self.parent_bead_timestamps.consensus_encode(w)?;
+        Ok(len)
+    }
+}
+
+impl Decodable for UnCommittedMetadata {
+    fn consensus_decode<R: BufRead + ?Sized>(r: &mut R) -> Result<Self, Error> {
+        let extra_nonce = i32::consensus_decode(r)?;
+        let broadcast_timestamp = Time::from_consensus(u32::consensus_decode(r).unwrap()).unwrap();
+        let signature = Signature::from_str(&String::consensus_decode(r).unwrap()).unwrap();
+        let parent_bead_timestamps = TimeVec::consensus_decode(r)?;
+        Ok(UnCommittedMetadata {
+            extra_nonce,
+            broadcast_timestamp,
+            signature,
+            parent_bead_timestamps,
+        })
+    }
+}
